reformat search functionality and create new ResultsPage comp
