+++ src/App.jsx
@@ -1,7 +1,7 @@
 // Library imports
 import React from "react";
-import { useState } from "react";
-import { Routes, Route } from "react-router-dom";
+import { useState, useEffect } from "react";
+import { Routes, Route, useLocation } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 
@@ -24,27 +24,16 @@ const queryClient = new QueryClient({
 });
 
 function App() {
-  // Init state
-  const [query, setQuery] = useState("");
-
-  // Update query value
-  function updateQueryValue(e) {
-    setQuery(e);
-  }
-
   return (
     <QueryClientProvider client={queryClient}>
       <Routes>
-        <Route
-          path="/"
-          element={<Navbar updateQueryValue={updateQueryValue} query={query} setQuery={setQuery} />}
-        >
+        <Route path="/" element={<Navbar />}>
           <Route index element={<Home />} />
           <Route path="movies" element={<Movies />} />
           <Route path="series" element={<Series />} />
           <Route path="details/movie/:id" element={<Details type="movie" />} />
           <Route path="details/tv/:id" element={<DetailsTV type="tv" />} />
-          <Route path="search" element={<ResultsPage query={query} />} />
+          <Route path="search" element={<ResultsPage />} />
         </Route>
       </Routes>
       <ReactQueryDevtools initialIsOpen={false} />
