+++ src/Components/CastRow.jsx
@@ -1,6 +1,6 @@
 function CastRow({ data, path }) {
-  const castRow = data?.map((item) => (
-    <div className="cast-item" key={item.id}>
+  const castRow = data?.map((item, index) => (
+    <div className="cast-item" key={index}>
       <img src={item.profile_path ? `${path}${item.profile_path}` : ""} />
       <span className="cast-name">{item.name}</span>
     </div>
